wip, breaks out components and builds timeline
