+++ src/components/GreetingSection/GreetingSection.module.css
@@ -1,0 +1,7 @@
+.greetingSection { }
+
+.greetingSection {
+  padding-bottom: 0;
+  padding-top: 16rem;
+  margin-bottom: 4rem;
+}
