+++ src/components/PersonalProjectsSection/PersonalProjectsSection.module.css
@@ -1,0 +1,48 @@
+.personalProjectsSection { }
+
+.personalProjectsSection a {
+  color: #656565;
+}
+
+.personalProjectsHeader {
+  font-size: 1.25rem;
+  margin-bottom: 2rem;
+}
+
+.personalProjectsHeader > p:first-of-type {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 2rem;
+}
+
+.personalProjectSubsection {
+  margin-bottom: 2rem;
+}
+
+.personalProjectSubsection > p:first-of-type {
+  font-weight: bold;
+  font-size: 1.5rem;
+}
+
+.personalProjectSubsection img {
+  max-width: 100vw;
+  margin: 2rem 0 0 -1rem;
+}
+
+.projectDescription a {
+  color: var(--colorBrandPrimary);
+}
+
+.projectHeader {
+  text-align: center;
+  margin-bottom: 2rem;
+}
+
+.projectHeader > p {
+  font-family: var(--fontFamilySerif);
+  font-size: 2rem;
+}
+
+.projectImgWrapper {
+  transform: translateY(calc(1rem * -1));
+}
