+++ src/components/PersonalProjectsSection/index.js
@@ -1,0 +1,1 @@
+export * from './PersonalProjectsSection';
