+++ src/components/ProfessionalSection/ProfessionalSection.module.css
@@ -1,0 +1,34 @@
+.professionalSection {
+  background: var(--colorBrandDark1);
+  color: #fff;
+  margin-left: -1rem;
+  padding: 2rem 0;
+  width: 100vw;
+}
+
+.professionalSection {
+  font-size: 1.125rem;
+  line-height: 1.25rem;
+  margin-bottom: 4rem;
+}
+
+.professionalRecent,
+.professionalPast {
+  line-height: 1.5rem;
+  padding: 0 1rem;
+}
+
+.professionalRecent {
+  margin-bottom: 2rem;
+}
+
+.professionalPast a {
+  color: #fff;
+}
+
+.professionalRecent__highlight {
+  font-size: 1.5rem;
+  font-weight: bold;
+  line-height: 2rem;
+  margin-bottom: 0.5rem;
+}
