+++ src/components/ProfesssionalTimeline/index.js
@@ -1,0 +1,1 @@
+export * from './ProfesssionalTimeline';
