+++ src/components/TextGreeting/TextGreeting.module.css
@@ -1,0 +1,12 @@
+.textGreeting { }
+
+.textGreeting {
+  font-family: var(--fontFamilySansSerif);
+  font-size: 1.5rem;
+}
+
+.bigText {
+  font-family: var(--fontFamilySerif);
+  font-weight: bold;
+  font-size: 4rem;
+}
